+++ findit/__init__.py
@@ -1,0 +1,1 @@
+from findit.core import FindIt
