+++ findit/engine.py
@@ -1,0 +1,14 @@
+class BaseEngine(object):
+    pass
+
+
+class EngineResult(object):
+    pass
+
+
+class TemplateEngine(BaseEngine):
+    pass
+
+
+class FeatureEngine(BaseEngine):
+    pass
