+++ findit/logger.py
@@ -1,0 +1,5 @@
+from loguru import logger
+
+# default: no log
+LOGGER_FLAG = 'findit'
+logger.disable(LOGGER_FLAG)
